change: move list product from navigation to product component
