+++ frontend/WebStoreAngular/src/app/app.component.html
@@ -2,4 +2,7 @@
     <app-navigation></app-navigation>
     <p>working</p>
     <h2>After navigation bar</h2>
+    <div>Before outlet</div>
+    <router-outlet></router-outlet>
+    <div>After outlet outlet</div>
 </div>
