+++ frontend/WebStoreAngular/src/app/models/category.ts
@@ -1,5 +1,3 @@
-//import { Subcategory } from "./subcategory";
-
 export interface Category {
   id: string;
   name: string;
