+++ frontend/WebStoreAngular/src/app/products/products.component.html
@@ -1,1 +1,6 @@
 <p>products works!</p>
+<div class="products">
+    <ul class="product" *ngFor="let product of getProducts">
+        {{product.name}}
+    </ul>
+</div>
