+++ frontend/WebStoreAngular/src/app/products/products.component.ts
@@ -1,5 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+import { Subscription } from 'rxjs';
 import { Product } from '../models/product';
+import { ShopService } from '../shop.service';
 
 @Component({
   selector: 'app-products',
@@ -8,7 +11,37 @@ import { Product } from '../models/product';
 })
 export class ProductsComponent implements OnInit {
   private subcategoryProducts: Product[] = [];
+  private sub: Subscription | undefined;
+  private products: Product[] = [];
 
-  constructor() {}
-  ngOnInit(): void {}
+  constructor(private route: ActivatedRoute, private shopService: ShopService) {
+    this.sub = this.shopService.products?.subscribe((product: Product[]) => {
+      this.products = product;
+      console.log(this.products);
+    });
+  }
+
+  ngOnInit(): void {
+    this.route.paramMap.subscribe((params) => {
+      const subcategoryId = params.get('id') as string;
+      console.log(subcategoryId);
+      if (subcategoryId) {
+        this.getProductsBySubcategoryId(subcategoryId);
+      }
+    });
+  }
+
+  private getProductsBySubcategoryId(subcategoryId: string) {
+    console.log('start gets products2');
+    this.shopService
+      .getProductsBySubcategory(subcategoryId)
+      .subscribe((products) => {
+        this.subcategoryProducts = products;
+        console.log(this.subcategoryProducts);
+      });
+  }
+
+  public get getProducts(): Product[] {
+    return this.subcategoryProducts;
+  }
 }
